Move impl_array_newtype to include/array_newtype.rs
The array_newtype.rs include file has been used for bitcoin's
previously internal macro impl_array_newtype_stringify. Since the name
lends itself to dual usage with the impl_array_newtype macro, and the
two may be used together, the macro can be copied from internals to
this include file.

Move impl_array_newtype macro to include/array_newtype.rs. Deprecate in
internals and replace uses with include! copy.
