+++ key_expression/src/lib.rs
@@ -25,7 +25,7 @@ extern crate hex;
 #[cfg(feature = "serde")]
 extern crate serde;
 
-// Pull in shared impl_array_newtype_stringify macro from include
+// Pull in shared macros from include
 // The impl_array_newtype_stringify requires crate::serde, $crate::hex and
 // crate::hashes to exist.
 #[cfg(feature = "alloc")]
