+++ p2p/src/lib.rs
@@ -8,6 +8,8 @@
 #![warn(deprecated_in_future)]
 #![doc(test(attr(warn(unused))))]
 
+include!("../include/array_newtype.rs");
+
 mod network_ext;
 
 #[cfg(feature = "std")]
